Basic collisions working.
Basic collisions working now. Using an invisible tile layer to act as a collision. Right now the center of the player is checked if it's in a collision tile, if so, move back to the last good position. Will add checking each bbox point of the player with tiles later.